Add paste image
Refactor and dedupe editor textarea code
