+++ src/features/post/new/NewPostText.tsx
@@ -1,4 +1,3 @@
-import { styled } from "@linaria/react";
 import {
   IonBackButton,
   IonButton,
@@ -8,46 +7,11 @@ import {
   IonTitle,
   IonToolbar,
 } from "@ionic/react";
-import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import { Centered, Spinner } from "../../auth/login/LoginNav";
-import TextareaAutosizedForOnScreenKeyboard from "../../shared/TextareaAutosizedForOnScreenKeyboard";
-import MarkdownToolbar, {
-  TOOLBAR_HEIGHT,
-  TOOLBAR_TARGET_ID,
-} from "../../shared/markdown/editing/MarkdownToolbar";
-import useKeyboardOpen from "../../../helpers/useKeyboardOpen";
-import useTextRecovery, {
-  clearRecoveredText,
-} from "../../../helpers/useTextRecovery";
-import { preventModalSwipeOnTextSelection } from "../../../helpers/ionic";
+import { clearRecoveredText } from "../../../helpers/useTextRecovery";
 import AppHeader from "../../shared/AppHeader";
-
-const Container = styled.div<{ keyboardOpen: boolean }>`
-  min-height: 100%;
-
-  display: flex;
-  flex-direction: column;
-
-  padding-bottom: ${({ keyboardOpen }) =>
-    keyboardOpen
-      ? TOOLBAR_HEIGHT
-      : `calc(${TOOLBAR_HEIGHT} + var(--ion-safe-area-bottom, env(safe-area-inset-bottom)))`};
-`;
-
-const Textarea = styled(TextareaAutosizedForOnScreenKeyboard)`
-  border: 0;
-  background: none;
-  resize: none;
-  outline: 0;
-  padding: 1rem;
-
-  flex: 1 0 auto;
-  min-height: 7rem;
-
-  html.ios:not(.theme-dark) & {
-    background: var(--ion-item-background);
-  }
-`;
+import Editor from "../../shared/markdown/editing/Editor";
 
 interface NewPostTextProps {
   value: string;
@@ -64,17 +28,12 @@ export default function NewPostText({
 }: NewPostTextProps) {
   const [loading, setLoading] = useState(false);
 
-  const keyboardOpen = useKeyboardOpen();
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
-
   const [text, setText] = useState(value);
   const isSubmitDisabled = loading;
 
   useEffect(() => {
     setValue(text);
   }, [setValue, text]);
-
-  useTextRecovery(text, setText, editing);
 
   async function submit() {
     if (isSubmitDisabled) return;
@@ -115,28 +74,11 @@ export default function NewPostText({
         </IonToolbar>
       </AppHeader>
       <IonContent>
-        <Container keyboardOpen={keyboardOpen}>
-          <Textarea
-            {...preventModalSwipeOnTextSelection}
-            id={TOOLBAR_TARGET_ID}
-            ref={textareaRef}
-            value={text}
-            onInput={(e) => setText((e.target as HTMLInputElement).value)}
-            autoFocus
-            onKeyDown={(e) => {
-              if ((e.ctrlKey || e.metaKey) && e.key === "Enter") {
-                submit();
-              }
-            }}
-          />
-        </Container>
-
-        <MarkdownToolbar
-          slot="fixed"
-          type="post"
+        <Editor
           text={text}
           setText={setText}
-          textareaRef={textareaRef}
+          canRecoverText={!editing}
+          onSubmit={submit}
         />
       </IonContent>
     </>
